requirements.txt
pytest==7.4.2
allure-pytest==2.13.2
selene>=2.0.0rc4
python-dotenv~=1.0.0
selenium==4.12.0
